// src/pages/modules/Pulsikom/about/chief-officer/officially/data/types.ts
export interface IOfficially {
  id_pimpinan: string
  id_satuan_organisasi: string
  id_kelompok_pimpinan: string
  url_gambar: string
  key_gambar: string
  nama_penjabat: string
  jabatan: string
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
  nama_user_created: string
  nama_user_updated: string
  nama_kelompok_pimpinan: string
  id_sdm: string
  is_dosen: boolean
}